requirements.txt
asgiref==3.8.1
click==8.1.7
dj-database-url==2.1.0
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
h11==0.14.0
packaging==24.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.11.0
uvicorn==0.29.0
